requirements.txt
google-cloud-bigquery
mysql-connector